IT-86
Changed reservation form where parties of 8 or more are allowed to make a reservation. Will need to merge with Alder's branch to make sure data validation is working.

Also opened IT-96 for a bug when linking reservation button to landing page. Reservation needs to have a button that looks similar to order now.